First naive implementation
This commit includes a working naive coordinate evaluation algo, with
support for a minimum_confidence threshold. Also baked in at this point
is the very simple API approach of file upload + query param. Finally,
up.json and .upignore are included, along with necessary build.gradle
changes to allow for a successful deploy to lambda. As of this commit,
catfindr.kono.sh is accessible publicly.